+++ crates/glam/src/f32/wasm/vec4.rs
@@ -1,12 +1,15 @@
 // Generated from vec.rs.tera template. Edit the template, not the generated file.
 
-use crate::{f32::math, wasm32::*, BVec4, BVec4A, Vec2, Vec3, Vec3A};
+use crate::{f32::math, wasm::*, BVec4, BVec4A, Vec2, Vec3, Vec3A};
 
 use core::fmt;
 use core::iter::{Product, Sum};
 use core::{f32, ops::*};
 
+#[cfg(target_arch = "wasm32")]
 use core::arch::wasm32::*;
+#[cfg(target_arch = "wasm64")]
+use core::arch::wasm64::*;
 
 #[cfg(feature = "zerocopy")]
 use zerocopy_derive::*;
@@ -93,6 +96,8 @@ impl Vec4 {
     /// Vec4 uses Intel SSE2
     pub const USES_SSE2: bool = false;
     /// Vec4 uses WebAssembly 128-bit SIMD
+    pub const USES_WASM_SIMD: bool = true;
+    #[deprecated(since = "0.31.0", note = "Renamed to USES_WASM_SIMD")]
     pub const USES_WASM32_SIMD: bool = true;
 
     /// Creates a new vector.
@@ -1156,6 +1161,18 @@ impl Vec4 {
     #[must_use]
     pub fn as_u64vec4(self) -> crate::U64Vec4 {
         crate::U64Vec4::new(self.x as u64, self.y as u64, self.z as u64, self.w as u64)
+    }
+
+    /// Casts all elements of `self` to `isize`.
+    #[inline]
+    #[must_use]
+    pub fn as_isizevec4(self) -> crate::ISizeVec4 {
+        crate::ISizeVec4::new(
+            self.x as isize,
+            self.y as isize,
+            self.z as isize,
+            self.w as isize,
+        )
     }
 
     /// Casts all elements of `self` to `usize`.