+++ crates/glam/src/f64/dvec2.rs
@@ -80,6 +80,8 @@ impl DVec2 {
     /// DVec2 uses Intel SSE2
     pub const USES_SSE2: bool = false;
     /// DVec2 uses WebAssembly 128-bit SIMD
+    pub const USES_WASM_SIMD: bool = false;
+    #[deprecated(since = "0.31.0", note = "Renamed to USES_WASM_SIMD")]
     pub const USES_WASM32_SIMD: bool = false;
 
     /// Creates a new vector.
@@ -1139,6 +1141,13 @@ impl DVec2 {
     #[must_use]
     pub fn as_u64vec2(self) -> crate::U64Vec2 {
         crate::U64Vec2::new(self.x as u64, self.y as u64)
+    }
+
+    /// Casts all elements of `self` to `isize`.
+    #[inline]
+    #[must_use]
+    pub fn as_isizevec2(self) -> crate::ISizeVec2 {
+        crate::ISizeVec2::new(self.x as isize, self.y as isize)
     }
 
     /// Casts all elements of `self` to `usize`.